Speed up the progress by further restricting progress bar display freq.
The whole progress bar gets refreshed only in steps of the $advance var.
Remove the $percent variable from show_progress(). Display a percent
value if the $max variable is >= 0, because that is when we can compute
a percent. Also remove the -p option that controled the display of the
percentage (but weirdly).